refactor!: move template to pkg
