+++ src/pkg/template/refs_test.go
