+++ src/pkg/template/testdata/configmap/expected.yaml
