+++ src/pkg/template/testdata/configmap/input.yaml
