+++ src/pkg/template/testdata/constants/expected.yaml
