+++ src/pkg/template/testdata/constants/input.yaml
