+++ src/pkg/template/testdata/deployment/expected.yaml
