+++ src/pkg/template/testdata/deployment/input.yaml
