+++ src/pkg/template/testdata/invalid/input.yaml
