+++ src/pkg/template/testdata/sprig-basic/expected.yaml
