+++ src/pkg/template/testdata/sprig-basic/input.yaml
