+++ src/pkg/template/testdata/sprig-functions/expected.yaml
