+++ src/pkg/template/testdata/sprig-functions/input.yaml
